[Feat] delete 버튼 클릭시 이미지 업로드 input 초기화 #102
- useImperativeHandle 이용
